Test/*: Reorganized headers; corrected C89-style declarations. Constructions/KeccakDuplex.c: Added an explicit cast Modes/KeccakHash.c: Added an explicit cast

+++ Modes/KeccakHash.c
@@ -43,7 +43,7 @@ HashReturn Keccak_HashUpdate(Keccak_HashInstance *instance, const BitSequence *d
             // The last partial byte is assumed to be aligned on the least significant bits
             unsigned char lastByte = data[databitlen/8];
             // Concatenate the last few bits provided here with those of the suffix
-            unsigned short delimitedLastBytes = (unsigned short)lastByte | ((unsigned short)instance->delimitedSuffix << (databitlen % 8));
+            unsigned short delimitedLastBytes = (unsigned short)(lastByte | (instance->delimitedSuffix << (databitlen % 8)));
             if ((delimitedLastBytes & 0xFF00) == 0x0000) {
                 instance->delimitedSuffix = delimitedLastBytes & 0xFF;
             }
